Rename `string` module in artichoke-backend to `fmt`
This module exclusively contains types and utilities for bridging
`std::fmt` with Artichoke `Error`s.
