+++ artichoke-backend/src/extn/core/array/args.rs
@@ -2,7 +2,7 @@ use std::fmt::Write as _;
 
 use crate::convert::implicitly_convert_to_int;
 use crate::extn::prelude::*;
-use crate::string::WriteError;
+use crate::fmt::WriteError;
 use crate::sys::protect;
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
