+++ artichoke-backend/src/extn/core/integer/mod.rs
@@ -3,7 +3,7 @@ use std::mem;
 
 use crate::extn::core::numeric::{self, Coercion, Outcome};
 use crate::extn::prelude::*;
-use crate::string::WriteError;
+use crate::fmt::WriteError;
 
 pub mod mruby;
 pub mod trampoline;
