+++ artichoke-backend/src/lib.rs
@@ -124,6 +124,7 @@ mod eval;
 pub mod exception_handler;
 pub mod extn;
 pub mod ffi;
+pub mod fmt;
 pub mod gc;
 mod globals;
 mod hash;
@@ -142,7 +143,6 @@ mod prng;
 mod regexp;
 pub mod release_metadata;
 pub mod state;
-pub mod string;
 pub mod sys;
 mod top_self;
 pub mod types;
